ARM: dts: rockchip: Add workaround for sdmmc boot on rk3588
This adds a workaround in order to allow booting from sdmmc on RK3588.

Two of the sdmmc clocks is controlled by registers only accessible in
secure-world, and is normally controlled using SCMI running in TF-A.

This cause issues in U-Boot SPL because TF-A is not yet running and
clock nodes with a compatible handled by the scmi_smccc driver.

However, when the bootrom boots from sdmmc it sets all clocks required
for SPL to read from the sd-card. Take advantage of this and define a
fixed-clock and replace the two scmi_clk with this fixed-clock.

Also changes rockchip_dwmmc_get_mmc_clk to return 0 when clk_set_rate
fails. Otherwise dwmci_setup_bus continues as the return value (uint)-1
is the correct sclk rate. Using 0 will result in the bus continue in
bypass mode.

Signed-off-by: Jonas Karlman <jonas@kwiboo.se>
[Updated to latest mainline code, use rk3588s-u-boot-dtsi for clock changes.
Add u-boot,spl-fifo-mode to sdmmc node to fix loading of all images]
Signed-off-by: Liviu Dudau <liviu@dudau.co.uk>